+++ clownaudio/decoders/libopenmpt.c
@@ -12,11 +12,11 @@
 struct Decoder_libOpenMPT
 {
 	openmpt_module *module;
-	unsigned int sample_rate;
+	unsigned long sample_rate;
 	unsigned int channel_count;
 };
 
-Decoder_libOpenMPT* Decoder_libOpenMPT_Create(DecoderData *data, bool loops, unsigned int sample_rate, unsigned int channel_count, DecoderInfo *info)
+Decoder_libOpenMPT* Decoder_libOpenMPT_Create(DecoderData *data, bool loops, unsigned long sample_rate, unsigned int channel_count, DecoderInfo *info)
 {
 	Decoder_libOpenMPT *decoder = NULL;
